課題２：モジュールディレクトリ構成
